\name{FunctionalNetworks-internal}
\alias{bic.generation}
\alias{data.generation}
\alias{network.estimation}

\title{
Internal functions of the FunctionalNetworks package.
}
\description{
These are internal functions of the package FunctionalNetworks. Check package Vignette for example on the usage of these functions.
}
%\usage{ 
%
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%
%}
\details{
These functions perform internal tasks an work within the package FunctionalNetworks.
}
%\value{
%An entry of the vector with the posterior sample for the parameter Alpha for the cross-sectional data setting.
%}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{
A. Quiroz-Zarate.
\email{aquiroz@jimmy.harvard.edu}
}
\keyword{internal}
%%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
